added login mutation, user auth middleware, clientside queries and mutations

+++ server/schemas/resolvers.js
@@ -1,7 +1,7 @@
 const { AuthenticationError } = require('apollo-server-express');
 const { Post } = require('../models');
 const { signToken } = require('../utils/auth');
-
+require('dotenv').config();
 const resolvers = {
     Query: {
         post: async(parent, { _id }) => {
@@ -15,35 +15,45 @@ const resolvers = {
 
     Mutation: {
         login: async (parent, { username, password }) => {
-            const admin = process.env.USERNAME;
-            const pass = process.env.PASSWORD;
+            const admin = process.env.DB_USERNAME;
+            const pass = process.env.DB_PASSWORD;
 
             if(admin != username || pass != password) {
-                throw new AuthenticationError('Incorrect credentials');
+
+                throw new AuthenticationError(`Incorrect credentials ${admin} ${username}`);
             }
 
             const token = signToken(admin);
 
-            return { token, admin };
+            return { token };
         },
-        addPost: async (parent, args) => {
-            const post = await Post.create({...args});
+        addPost: async (parent, args, context) => {
+            if (context.admin) {
+                const post = await Post.create({...args});
 
-            return post ;
+                return post ;
+            }
+            throw new AuthenticationError('Not logged in!');
         },
-        editPost: async (parent, args) => {
-            return Post.findOneAndUpdate(
-                { _id: args._id},
-                { header: args.header, body: args.body, video: args.video, category: args.category },
-                { new: true }
-            )
+        editPost: async (parent, args, context) => {
+            if (context.admin) {
+                return Post.findOneAndUpdate(
+                    { _id: args._id},
+                    { header: args.header, body: args.body, video: args.video, category: args.category },
+                    { new: true }
+                )
+            }
+            throw new AuthenticationError('Not logged in!');
         },
-        removePost: async (parent, { _id }) => {
-            const deletedThought = await Post.findOneAndDelete(
-                { _id: _id },
-                { new: true }
-            );
-            return deletedThought;
+        removePost: async (parent, { _id }, context) => {
+            if (context.admin) {
+                const deletedThought = await Post.findOneAndDelete(
+                    { _id: _id },
+                    { new: true }
+                );
+                return deletedThought;
+            }
+            throw new AuthenticationError('Not logged in!');
         }
 
     }
